CLOUDP-118864: Split cloud manager and atlas decryptions
