+++ e2e/atlas/decryption/aws/test-input
