+++ e2e/atlas/decryption/gcp/test-input
