+++ e2e/atlas/decryption/gcp/test-output
