+++ e2e/atlas/decryption_aws_test.go
@@ -11,56 +11,47 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-//go:build e2e || (decrypt && (cloudmanager || om44 || om50))
+//go:build e2e || (atlas && decrypt)
 
-package cloud_manager_test
+package atlas_test
 
 import (
 	"embed"
-	"io/fs"
 	"os"
 	"os/exec"
-	"path"
 	"testing"
 
 	"github.com/mongodb/mongocli/e2e"
+	"github.com/mongodb/mongocli/e2e/decryption"
 	"github.com/stretchr/testify/require"
 )
 
-//go:embed decryption/gcp/*
-var filesGCP embed.FS
+//go:embed decryption/aws/*
+var filesAWS embed.FS
 
-const gcpTestsInputDir = "decryption/gcp"
+const awsTestsInputDir = "decryption/aws"
 
-func TestDecryptWithGCP(t *testing.T) {
+func TestDecryptWithAWS(t *testing.T) {
 	req := require.New(t)
 
-	cliPath, err := e2e.Bin()
+	cliPath, err := e2e.AtlasCLIBin()
 	req.NoError(err)
 
 	tmpDir := t.TempDir()
-
-	GCPCredentialsContent, ok := os.LookupEnv("GCP_CREDENTIALS")
-	req.True(ok, "GCP Credentials not found")
-	GCPCredentialsFile := path.Join(tmpDir, "gcp_credentials.json")
-	err = os.WriteFile(GCPCredentialsFile, []byte(GCPCredentialsContent), fs.ModePerm)
-	req.NoError(err)
-	t.Setenv("GOOGLE_APPLICATION_CREDENTIALS", GCPCredentialsFile)
 
 	t.Cleanup(func() {
 		err = os.RemoveAll(tmpDir)
 		req.NoError(err)
 	})
 
-	inputFile := generateFileName(tmpDir, "input")
-	err = dumpToTemp(filesGCP, generateFileName(gcpTestsInputDir, "input"), inputFile)
+	inputFile := decryption.GenerateFileName(tmpDir, "input")
+	err = decryption.DumpToTemp(filesAWS, decryption.GenerateFileName(awsTestsInputDir, "input"), inputFile)
 	req.NoError(err)
 
-	expectedContents, err := filesGCP.ReadFile(generateFileName(gcpTestsInputDir, "output"))
+	expectedContents, err := filesAWS.ReadFile(decryption.GenerateFileName(awsTestsInputDir, "output"))
 	req.NoError(err)
 
 	cmd := exec.Command(cliPath,
-		entity,
 		"logs",
 		"decrypt",
 		"--file",
@@ -71,7 +62,7 @@ func TestDecryptWithGCP(t *testing.T) {
 	gotContents, err := cmd.CombinedOutput()
 	req.NoError(err, string(gotContents))
 
-	equal, err := logsAreEqual(expectedContents, gotContents)
+	equal, err := decryption.LogsAreEqual(expectedContents, gotContents)
 	req.NoError(err)
 	req.True(equal, "expected %v, got %v", string(expectedContents), string(gotContents))
 }
