+++ e2e/cloud_manager/decryption_kmip_test.go
@@ -27,6 +27,7 @@ import (
 	"testing"
 
 	"github.com/mongodb/mongocli/e2e"
+	"github.com/mongodb/mongocli/e2e/decryption"
 	"github.com/stretchr/testify/require"
 )
 
@@ -80,11 +81,11 @@ func TestDecryptWithKMIP(t *testing.T) {
 
 	for i := 1; i <= 2; i++ {
 		t.Run(fmt.Sprintf("Test case %v", i), func(t *testing.T) {
-			inputFile := generateFileNameCase(tmpDir, i, "input")
-			err := dumpToTemp(filesKmip, generateFileNameCase(kmipTestsInputDir, i, "input"), inputFile)
+			inputFile := decryption.GenerateFileNameCase(tmpDir, i, "input")
+			err := decryption.DumpToTemp(filesKmip, decryption.GenerateFileNameCase(kmipTestsInputDir, i, "input"), inputFile)
 			req.NoError(err)
 
-			expectedContents, err := filesKmip.ReadFile(generateFileNameCase(kmipTestsInputDir, i, "output"))
+			expectedContents, err := filesKmip.ReadFile(decryption.GenerateFileNameCase(kmipTestsInputDir, i, "output"))
 			req.NoError(err)
 
 			cmd := exec.Command(cliPath,
@@ -103,7 +104,7 @@ func TestDecryptWithKMIP(t *testing.T) {
 			gotContents, err := cmd.CombinedOutput()
 			req.NoError(err, string(gotContents))
 
-			equal, err := logsAreEqual(expectedContents, gotContents)
+			equal, err := decryption.LogsAreEqual(expectedContents, gotContents)
 			req.NoError(err)
 			req.True(equal, "expected %v, got %v", string(expectedContents), string(gotContents))
 		})
